Refactoring on base objects (#7)
* Added test case for missing filter

* Changed FIELD constant prefix to ATTRIBUTE

To avoid confusion with Fields that are part of the YAML Schema, rename field names to ATTRIBUTE_

* Renamed export to snapshot

Export was a confusing name to get the current state of the object tree.  So, named it snapshot.

* Moved nested schema loaders to base schema collection

To reduce repetitions in the subclasses, created a SchemaCollecteion base object that loads nested schema

* Moved nested item iniitalization to the base class

All sub-items as defined in the schema is initialized in the base Item Collection now

* Improved comments

* Added the capacity to create and restore snapshots

Snapshot and restore helps manage existing state

* Refactored and improved the DataGroup code.

* made nested schema and nested items consistent across base and collection definitions

* Removed evaluation context in the base

To ensure greater transparency context objects have to be passed from higher up the stack... there is no implicit context generation at the base, which could lead to untracable issues.

* Variable naming of Context

* Cleaned up session datagroup schema

Refactored to follow the base inheritece pattern

* Cleaned up session data group

* Cleaned up and documented exceptions

* Fixed some issues with context passing

All evaluations must be passed the context to evaluate with

* Added streaming trasnsformer as the subclass of Transformer

To conform to the design pattern, converted Transformer to an abstract base from which the Streaming Transformer is inherited.

* Made field ABC to conform to pattern

Since we have an abstract basea schema and abstract base for each stage,  I created a simple field that inherits from the abstract field - to maintain consistency with the rest.