All base objects are created - foundation is ready. (#2)
* Base Field Schema and Field store

* Added Data Group

Data Group & Data Group Schema first iteration.

* Added Transform Schema and renamed Data Groups to Groups

Trying to keep the names lean.

* Added Session Group and Transformer

* Changed filter and value to use compiled expressions. Also structured the code to use some base classes.

* Major refactor to shrink the number of files

Added store and did mass renaming of files.

* Consolidated schema into one file

* Restored Field that was accidentally deleted during merge

* Added unit tests

* Fix record, filter_expr and take in a global and local context in the Interpreter.

* Added more base tests

* Restructure project to move python environment to root

Changed the reference paths to ensure that full module path is useed for cross references.

* Cleaned up base schema

* Added a running context and fixed issues in getting expression evaluation working.

* Changed field to workok with the new base

* Added Expression tests to completion

Changed Expression to contain execution.

* Fixed session predefined fields to use the name of the data group.

* Added a sub_item property to BaseItem. Also added support to build json from the item objects.

* Added Types

Type definitions for different all the types supported by the system

* Execution result and execution fixing

Imporoved execution by adding execution result and error tracking

* Comleted unit tests for BaseItem

Split BaseItemGroup out of BaseItem to represent collections

* Removed EvaluationResult

Added unit test cases for base

* Renaming to Data Gropu

Renamed Group to Data Group
Renamed Base Item Group to BaseItemCollection

* Renamed SessionDataGroup

* Added unit tests for expression

* Dynamic type loading from type string

* Fixed transfer context referec

Now the runner is executing without errors

* Update running to not create a new transformer object when executing.

* Updated loader to distinguish between item and schema

* Refactoring to add types everywhere

* Added local runner with sample data to tests. The code is able to evaluate a simple yaml file and generate results.

* YAPF'ed the code

* Refactoring on base objects (#7)

* Added test case for missing filter

* Changed FIELD constant prefix to ATTRIBUTE

To avoid confusion with Fields that are part of the YAML Schema, rename field names to ATTRIBUTE_

* Renamed export to snapshot

Export was a confusing name to get the current state of the object tree.  So, named it snapshot.

* Moved nested schema loaders to base schema collection

To reduce repetitions in the subclasses, created a SchemaCollecteion base object that loads nested schema

* Moved nested item iniitalization to the base class

All sub-items as defined in the schema is initialized in the base Item Collection now

* Improved comments

* Added the capacity to create and restore snapshots

Snapshot and restore helps manage existing state

* Refactored and improved the DataGroup code.

* made nested schema and nested items consistent across base and collection definitions

* Removed evaluation context in the base

To ensure greater transparency context objects have to be passed from higher up the stack... there is no implicit context generation at the base, which could lead to untracable issues.

* Variable naming of Context

* Cleaned up session datagroup schema

Refactored to follow the base inheritece pattern

* Cleaned up session data group

* Cleaned up and documented exceptions

* Fixed some issues with context passing

All evaluations must be passed the context to evaluate with

* Added streaming trasnsformer as the subclass of Transformer

To conform to the design pattern, converted Transformer to an abstract base from which the Streaming Transformer is inherited.

* Made field ABC to conform to pattern

Since we have an abstract basea schema and abstract base for each stage,  I created a simple field that inherits from the abstract field - to maintain consistency with the rest.

* Added very basic session split support.

* Change _dict_store back to _cache in the base store. Was changed by mistake earlier.

* Transfromer and Execution Context Cleanup (#14)

* Added test case for missing filter

* Changed FIELD constant prefix to ATTRIBUTE

To avoid confusion with Fields that are part of the YAML Schema, rename field names to ATTRIBUTE_

* Renamed export to snapshot

Export was a confusing name to get the current state of the object tree.  So, named it snapshot.

* Moved nested schema loaders to base schema collection

To reduce repetitions in the subclasses, created a SchemaCollecteion base object that loads nested schema

* Moved nested item iniitalization to the base class

All sub-items as defined in the schema is initialized in the base Item Collection now

* Improved comments

* Added the capacity to create and restore snapshots

Snapshot and restore helps manage existing state

* Refactored and improved the DataGroup code.

* made nested schema and nested items consistent across base and collection definitions

* Removed evaluation context in the base

To ensure greater transparency context objects have to be passed from higher up the stack... there is no implicit context generation at the base, which could lead to untracable issues.

* Variable naming of Context

* Cleaned up session datagroup schema

Refactored to follow the base inheritece pattern

* Cleaned up session data group

* Cleaned up and documented exceptions

* Fixed some issues with context passing

All evaluations must be passed the context to evaluate with

* Added streaming trasnsformer as the subclass of Transformer

To conform to the design pattern, converted Transformer to an abstract base from which the Streaming Transformer is inherited.

* Made field ABC to conform to pattern

Since we have an abstract basea schema and abstract base for each stage,  I created a simple field that inherits from the abstract field - to maintain consistency with the rest.

* Changed contexts to execution context for better flow of data

* Added Evaluation Context to everything

* YAPF Formatting

* PR addressed